python/requirements.txt
################################################################
#
# Python package requirements for the TeamClear:
# https://github.com/taking/TeamClear
#
#     pip install -r requirements.txt
#
################################################################

flask
flask_bootstrap
simplejson
pillow

################################################################
#
#
# flask - Server (Web)
# flask_bootstrap
# simplejson
# pillow
#
#
################################################################
